# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Inventory System API",
    "description": "API documentation for the Inventory System"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Registers a new user with a username, password, and type",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": ["student", "admin"]
                }
              },
              "required": ["username", "password", "type"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login a user",
        "description": "Logs in a user and returns a JWT token",
        "parameters": [
          {
            "in": "body",
            "name": "credentials",
            "description": "User credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/items": {
      "post": {
        "summary": "Add a new item",
        "description": "Adds a new item to the inventory",
        "parameters": [
          {
            "in": "body",
            "name": "item",
            "description": "Item object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer"
                },
                "description": {
                  "type": "string"
                }
              },
              "required": ["name", "quantity", "description"]
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "JWT token"
          }
        ],
        "responses": {
          "201": {
            "description": "Item created"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "get": {
        "summary": "Get all items",
        "description": "Retrieves all items from the inventory",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "JWT token"
          }
        ],
        "responses": {
          "200": {
            "description": "List of items",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/items/{id}": {
      "delete": {
        "summary": "Delete an item",
        "description": "Deletes an item from the inventory",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Item ID"
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "JWT token"
          }
        ],
        "responses": {
          "204": {
            "description": "Item deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
    "/approve-order/{id}": {
      "post": {
        "summary": "Approve an order",
        "description": "Approves an order and creates a loan record",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "Order ID"
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "JWT token"
          }
        ],
        "responses": {
          "200": {
            "description": "Order approved"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Order not found"
          },
          "400": {
            "description": "Item not available"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": ["student", "admin"]
        }
      },
      "required": ["username", "password", "type"]
    },
    "Item": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "quantity": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        }
      },
      "required": ["name", "quantity", "description"]
    },
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "userId": {
          "type": "integer"
        },
        "itemId": {
          "type": "integer"
        }
      },
      "required": ["userId", "itemId"]
    },
    "CurrentLoan": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "userId": {
          "type": "integer"
        },
        "itemId": {
          "type": "integer"
        },
        "loanDate": {
          "type": "string",
          "format": "date-time"
        },
        "returnDate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": ["userId", "itemId", "loanDate"]
    }
  }
}